html/semantics/embedded-content/media-elements/audio_currentSrc_src_null.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Audio Test: audio_currentSrc_src_null</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2014/CR-html5-20140429/embedded-content-0.html#htmlmediaelement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<audio id="m">
  Your browser does not support media elements.
</audio>
<div id="log"></div>
<script>
test(function () {
  var media = document.getElementById("m");
  media.src = null;
  assert_equals(media.currentSrc, '', "The return value of audio.currentSrc");
}, "Check if the audio.currentSrc returns the empty string when there is no media resource");
</script>
